+++ packages/web/src/app/doc-pages/about/the-design-system-doc/flexible-icon/flexible-icon.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-flexible-icon',
+  templateUrl: './flexible-icon.component.html',
+})
+export class FlexibleIconComponent {}