+++ packages/web/src/app/doc-pages/about/the-design-system-doc/quality-icon/quality-icon.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-quality-icon',
+  templateUrl: './quality-icon.component.html',
+})
+export class QualityIconComponent {}