+++ packages/web/src/app/home/home.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { Title } from '@angular/platform-browser';
 import { LOCALE_CODE } from 'contentful/types';
@@ -7,11 +7,14 @@ import { homeMenu } from 'src/app/shared/doc-pages';
 import changelogJson from '@elvia/elvis/CHANGELOG.json';
 import { Observable } from 'rxjs';
 import { BreakpointService } from '../core/services/breakpoint.service';
+import { ThemeService } from '../core/services/theme.service';
+import { ThemeName } from '@elvia/elvis-colors';
 
 @Component({
   selector: 'app-home',
   templateUrl: './home.component.html',
   styleUrls: ['./home.component.scss'],
+  encapsulation: ViewEncapsulation.None,
 })
 export class HomeComponent implements OnInit {
   overviewTitle = 'Elvia design system';
@@ -26,12 +29,14 @@ export class HomeComponent implements OnInit {
   private currentDate = new Date();
   currentYear = this.currentDate.getFullYear();
   locale: LOCALE_CODE;
+  currentTheme: ThemeName = 'light';
   changelog = changelogJson.content;
   isMobileScreenWidth: Observable<boolean>;
 
   constructor(
     localizationService: LocalizationService,
     breakpointService: BreakpointService,
+    private themeService: ThemeService,
     private titleService: Title,
   ) {
     this.isMobileScreenWidth = breakpointService.matches(['xs']);
@@ -47,6 +52,12 @@ export class HomeComponent implements OnInit {
           this.overviewTitle = 'Elvia designsystem';
           this.locale = 'nb-NO';
         }
+      });
+    this.themeService
+      .listenTheme()
+      .pipe(takeUntilDestroyed())
+      .subscribe((theme) => {
+        this.currentTheme = theme;
       });
   }