+++ packages/web/src/app/shared/feedback/emoji-neutral/emoji-neutral.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-emoji-neutral',
+  templateUrl: './emoji-neutral.component.html',
+})
+export class EmojiNeutralComponent {}