+++ packages/elvis-assets-icons/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Elvis-assets-icon
 
+## 3.4.0 (25.08.23)
+
+### New feature
+
+- Added the `compass` and `ruler` icons.
+
+## 3.3.1 (24.08.23)
+
+### Patch
+
+- Updated the css-variables for icon colors to new names.
+
 ## 3.3.0 (20.07.23)
 
 ### New feature